map spark Row to SparkRowAdapter before collection
